feat: add services

+++ internal/models/model_services.go
@@ -1,7 +1,7 @@
 package models
 
 type Services struct {
-	ServicesId string `gorm:"primaryKey" json:"servicesId"`
-	Name       string `json:"name"`
-	Price      int    `json:"price"`
+	ServicesId string  `gorm:"primaryKey" json:"servicesId"`
+	Name       string  `json:"name"`
+	Price      float64 `json:"price"`
 }
